Compute an outpath depending on a hash of the headers
The key value depends on the local computer where the mails are retrieved
and the datetime of retrieval. This makes sharing the same detachment folder
accross computers result in a lot of duplicate file. Using a more static
information, like the header of the file, allows to get a more stable output
path and ease sharing between computers.